// typechain-types/Consideration.ts
export type AssetStruct = {
  assetType: BigNumberish;
  token: string;
  identifier: BigNumberish;
  amount: BigNumberish;
};

export type AssetStructOutput = [number, string, BigNumber, BigNumber] & {
  assetType: number;
  token: string;
  identifier: BigNumber;
  amount: BigNumber;
};

export type ReceivedAssetStruct = {
  assetType: BigNumberish;
  token: string;
  identifier: BigNumberish;
  amount: BigNumberish;
  account: string;
};

export type ReceivedAssetStructOutput = [
  number,
  string,
  BigNumber,
  BigNumber,
  string
] & {
  assetType: number;
  token: string;
  identifier: BigNumber;
  amount: BigNumber;
  account: string;
};

export type OrderParametersStruct = {
  offer: AssetStruct[];
  consideration: ReceivedAssetStruct[];
  startTime: BigNumberish;
  endTime: BigNumberish;
  offerer: string;
  salt: BigNumberish;
  facilitator: string;
  nonce: BigNumberish;
};